refactor and organize files
